Configuration tree demo
This commit showcases how to mount a configuration tree. The feature is
only enabled when running on Kubernetes so it will not make any
difference when running this from an IDE.

To see the impact, the "active.on-cloud-platform" can be commented out
which likely leads to an application startup failure as the directory
does not exist. Here is a short script to create an override for the
"custom.test.property":

mkdir -p /tmp/test-config/custom/test
echo Hello >> /tmp/test-config/custom/test/property

Rather than enabling the `k8s` profile we import the additional
configuration file using "spring.config.import".